requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pandas==2.2.3
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.8.1
